Queue handler - first steps

+++ config/config.php
@@ -16,7 +16,7 @@ return [
     ],
     'rabbitmq' => [
         'host' => env('RABBITMQ_HOST'),
-        'port' => (int) env('RABBITMQ_PORT'),
+        'port' => env('RABBITMQ_PORT'),
         'user' => env('RABBITMQ_USER'),
         'password' => env('RABBITMQ_PASSWORD'),
     ],
